Separated mongoose bite from the cobra slayer rune, mongoose bite can only be used when defensive state aura is active, added threat multipliers
