+++ sim/hunter/wing_clip.go
@@ -35,6 +35,7 @@ func (hunter *Hunter) getWingClipConfig(rank int) core.SpellConfig {
 
 		CritDamageBonus: hunter.mortalShots(),
 		DamageMultiplier: 1,
+		ThreatMultiplier: 1,
 
 		ApplyEffects: func(sim *core.Simulation, target *core.Unit, spell *core.Spell) {
 			spell.CalcAndDealDamage(sim, target, baseDamage, spell.OutcomeMeleeWeaponSpecialHitAndCrit)
